fix: hooks 추가로 컴포넌트 재사용 가능하도록 수정
