+++ frontend/src/components/home/SolutionList.tsx
@@ -1,30 +1,17 @@
-import React, { useEffect, useState } from "react";
-
 import { Container, List, ListItem, ListItemText } from "@mui/material";
-import { solutionList } from "../../apis/v1/solutions";
-import { Solution } from "../../types/solution/Solution";
 import styles from "../../styles/SolutionList.module.css";
+import { Solution, SolutionList } from "../../types/solution/Solution";
 
-const CurrentList = () => {
-  const [solution_list, setSolutionList] = useState<Solution[] | null>(null);
-
-  useEffect(() => {
-    const initSolutionList = async () => {
-      const list = await solutionList();
-      setSolutionList(list || null);
-    };
-    initSolutionList();
-  }, []);
-
+const SolutionItemList = (props: any) => {
   return (
     <div>
       <Container>
         <div className={styles.title}>최근 문제 풀이</div>
-        {solution_list === null ? (
+        {props.list === null ? (
           <div className={styles.blank_list}>최근 문제 풀이가 없습니다.</div>
         ) : (
           <List className={styles.list_container}>
-            {solution_list?.map((solution) => (
+            {props?.list.map((solution: Solution) => (
               <ListItem
                 key={solution.id}
                 disableGutters
@@ -48,4 +35,4 @@ const CurrentList = () => {
   );
 };
 
-export default CurrentList;
+export default SolutionItemList;
