+++ frontend/src/pages/Home.tsx
@@ -1,7 +1,9 @@
+import { useEffect, useState } from "react";
 import NavBar from "../components/common/Nav";
 import Footer from "../components/common/Footer";
 import Banner from "../components/common/Banner";
 import SolutionList from "../components/home/SolutionList";
+import useSolutionList from "../hooks/solutionList";
 
 const Home = () => {
   return (
@@ -9,7 +11,7 @@ const Home = () => {
       <div>
         <NavBar />
         <Banner />
-        <SolutionList />
+        <SolutionList list={useSolutionList()} />
       </div>
       <div>
         <Footer />
