+++ frontend/src/types/solution/Solution.ts
@@ -1,11 +1,11 @@
 export interface Solution {
-    id: number;
-    title: string;
-    provider: string;
-    number: number;
-    username: string;
+  id: number;
+  title: string;
+  provider: string;
+  number: number;
+  username: string;
 }
 
 export interface SolutionList {
-    items: Solution[];
+  items: Solution[];
 }
